+++ src/main/java/com/tringles/tutorial/config/OAuth2/OAuth2SuccessHandler.java
@@ -1,0 +1,2 @@
+package com.tringles.tutorial.config.OAuth2;
+
